third_party/blink/web_tests/http/tests/eventsource/eventsource-reconnect-expected.txt

Test EventSource reconnect after end of event stream.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS es.readyState is es.CONNECTING
PASS es.readyState is es.OPEN
PASS evt.lastEventId is "77☃"
PASS es.readyState is es.CONNECTING
PASS es.readyState is es.OPEN
PASS evt.data is "77☃"
PASS evt.lastEventId is "77☃"
PASS es.readyState is es.CLOSED
PASS successfullyParsed is true

TEST COMPLETE
